+++ packages/analyzer/fixtures/class-attributes/package/my-element.js
@@ -5,6 +5,7 @@
 class MyEl extends HTMLElement {
   /**
    * @attr
+   * @reflect
    */
   myField;