+++ packages/analyzer/fixtures/class-fields/package/my-element.js
@@ -18,7 +18,8 @@ class MyEl extends HTMLElement {
 
   /**
    * This is also an attribute
-   * @attr
+   * @attr also-attr
+   * @reflect
    * @type {string}
    */
   alsoAttr;