+++ packages/analyzer/src/features/framework-plugins/lit/property-decorator.js
@@ -1,6 +1,6 @@
 import { decorator } from '../../../utils/index.js';
 import { createAttributeFromField } from '../../analyse-phase/creators/createAttribute.js';
-import { hasPropertyDecorator, isAlsoAttribute, getAttributeName } from './utils.js';
+import { hasPropertyDecorator, isAlsoAttribute, getAttributeName, reflects } from './utils.js';
 
 /**
  * PROPERTY
@@ -39,6 +39,15 @@ export function propertyDecoratorPlugin() {
                 const attributeName = getAttributeName(propertyOptions);
                 if(attributeName) {
                   attribute.name = attributeName;
+                  field.attribute = attributeName;
+                } else {
+                  field.attribute = field.name;
+                }
+
+                
+                if(reflects(propertyOptions)) {
+                  field.attribute = attribute.name;
+                  field.reflects = true;
                 }
 
                 const existingAttribute = currClass?.attributes?.find(attr => attr.name === attribute.name);