+++ packages/analyzer/src/features/framework-plugins/lit/static-properties.js
@@ -1,7 +1,7 @@
 import { createAttributeFromField } from '../../analyse-phase/creators/createAttribute.js';
 import { getDefaultValuesFromConstructorVisitor } from '../../analyse-phase/creators/createClass.js';
 import { handleJsDoc } from '../../analyse-phase/creators/handlers.js';
-import { isAlsoAttribute, hasStaticKeyword, getPropertiesObject, getAttributeName } from './utils.js';
+import { isAlsoAttribute, hasStaticKeyword, getPropertiesObject, getAttributeName, reflects } from './utils.js';
 
 /**
  * STATIC-PROPERTIES
@@ -41,6 +41,14 @@ export function staticPropertiesPlugin() {
                   const attributeName = getAttributeName(property);
                   if(attributeName) {
                     attribute.name = attributeName;
+                    classMember.attribute = attributeName;
+                  } else {
+                    classMember.attribute = classMember.name;
+                  }
+
+                  if(reflects(property)) {
+                    classMember.attribute = attribute.name;
+                    classMember.reflects = true;
                   }
 
                   currClass.attributes = [...(currClass?.attributes || []), attribute]